Added support for emacs prog-mode compilation
There are mainly two features in this commit. First, when doing
`M-x compile`, the suggested compilation command is now
`syncon-parser <filename>` (technically the buffer name, but close
enough). Second, running `M-x next-error` will now jump to the
position where the next error is listed. This works both for
errors in the syncon definition and the file being parsed by a
syncon definition.